+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added `Arena::invalidate` for invalidating indices on-demand, as a faster remove-followed-by-reinsert.
 * Added `Index::to_bits` and `Index::from_bits` for converting indices to a form convenient for passing outside of Rust.
 * Added `Arena::clear` for conveniently clearing the whole arena.
+* Change the semantics of `Arena::drain` to drop any remaining uniterated items when the `Drain` iterator is dropped, clearing the `Arena`.
 
 ## 0.2.1 (2020-10-01)
 * Added `Default` implementation for `Arena`.